requirements.txt
asn1crypto==0.24.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
cryptography==2.4.2
future==0.17.1
idna==2.7
pkg-resources==0.0.0
pycparser==2.19
python-telegram-bot==11.1.0
requests==2.20.1
six==1.11.0
urllib3==1.24.1
